apt-defender-system/pi-agent/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.25
aiosqlite==0.19.0
httpx==0.26.0
pyyaml==6.0.1
yara-python==4.5.0
psutil==5.9.8
supabase==2.3.08
py-cpuinfo==9.0.0
watchdog==4.0.0
cryptography==42.0.0
python-dotenv==1.0.0
requests==2.31.0
numpy==1.26.3
pandas==2.2.0
scikit-learn==1.4.0
